cambios en models guardado de receta y detalle receta usando attach

+++ database/migrations/2016_11_29_003533_create_medicamentos_has_recetas_table.php
@@ -16,7 +16,7 @@ class CreateMedicamentosHasRecetasTable extends Migration
           $table->primary(['medicamentos_id', 'recetas_id']);
           $table->integer('medicamentos_id')->unsigned()->index();
           $table->integer('recetas_id')->unsigned()->index();
-          $table->integer('dosis');
+          $table->float('dosis');
           $table->integer('dias');
           $table->integer('cantidad');
           $table->integer('periodicidad');
